+++ src/pages/Explore/TraceExploration.tsx
@@ -1,7 +1,7 @@
 import { css } from '@emotion/css';
 import React from 'react';
 
-import { AdHocVariableFilter, GrafanaTheme2 } from '@grafana/data';
+import { AdHocVariableFilter, GrafanaTheme2, LoadingState, PluginExtensionLink } from '@grafana/data';
 import {
   AdHocFiltersVariable,
   CustomVariable,
@@ -13,13 +13,14 @@ import {
   SceneObjectState,
   SceneObjectUrlSyncConfig,
   SceneObjectUrlValues,
+  SceneQueryRunner,
   SceneRefreshPicker,
   SceneTimePicker,
   SceneTimeRange,
   SceneVariableSet,
 } from '@grafana/scenes';
 import { config } from '@grafana/runtime';
-import { Badge, Button, Drawer, Dropdown, Icon, Menu, Stack, Tooltip, useStyles2 } from '@grafana/ui';
+import { Badge, Button, Drawer, Dropdown, Icon, IconButton, Menu, Stack, Tooltip, useStyles2 } from '@grafana/ui';
 
 import { TracesByServiceScene } from '../../components/Explore/TracesByService/TracesByServiceScene';
 import {
@@ -36,7 +37,7 @@ import {
   VAR_PRIMARY_SIGNAL,
   VAR_SPAN_LIST_COLUMNS,
 } from '../../utils/shared';
-import { getTraceExplorationScene, getFiltersVariable, getPrimarySignalVariable } from '../../utils/utils';
+import { getTraceExplorationScene, getFiltersVariable, getPrimarySignalVariable, getDataSource } from '../../utils/utils';
 import { TraceDrawerScene } from '../../components/Explore/TracesByService/TraceDrawerScene';
 import { primarySignalOptions } from './primary-signals';
 import { VariableHide } from '@grafana/schema';
@@ -44,6 +45,8 @@ import { reportAppInteraction, USER_EVENTS_ACTIONS, USER_EVENTS_PAGES } from 'ut
 import { PrimarySignalVariable } from './PrimarySignalVariable';
 import { renderTraceQLLabelFilters } from 'utils/filters-renderer';
 import { TraceQLIssueDetector, TraceQLConfigWarning } from '../../components/Explore/TraceQLIssueDetector';
+import { AddToInvestigationButton } from 'components/Explore/actions/AddToInvestigationButton';
+import { ADD_TO_INVESTIGATION_MENU_TEXT, getInvestigationLink } from 'components/Explore/panels/PanelMenu';
 
 export interface TraceExplorationState extends SceneObjectState {
   topScene?: SceneObject;
@@ -62,6 +65,9 @@ export interface TraceExplorationState extends SceneObjectState {
   initialFilters?: AdHocVariableFilter[];
 
   issueDetector?: TraceQLIssueDetector;
+
+  investigationLink?: PluginExtensionLink;
+  addToInvestigationButton?: AddToInvestigationButton;
 }
 
 const version = process.env.VERSION;
@@ -93,9 +99,14 @@ export class TraceExploration extends SceneObjectBase<TraceExplorationState> {
 
     this._subs.add(
       this.subscribeToEvent(EventTraceOpened, (event) => {
+        this.setupInvestigationButton(event.payload.traceId);
         this.setState({ traceId: event.payload.traceId, spanId: event.payload.spanId });
       })
     );
+
+    if (this.state.traceId) {
+      this.setupInvestigationButton(this.state.traceId);
+    }
 
     const datasourceVar = sceneGraph.lookupVariable(VAR_DATASOURCE, this) as DataSourceVariable;
     datasourceVar.subscribeToState((newState) => {
@@ -156,6 +167,69 @@ export class TraceExploration extends SceneObjectBase<TraceExplorationState> {
     this.setState({ traceId: undefined, spanId: undefined });
   }
 
+  private setupInvestigationButton(traceId: string) {
+    const traceExploration = getTraceExplorationScene(this);
+    const dsUid = getDataSource(traceExploration);
+
+    const queryRunner = new SceneQueryRunner({
+      datasource: { uid: dsUid },
+      queries: [{ 
+        refId: 'A', 
+        query: traceId, 
+        queryType: 'traceql',
+      }],
+    });
+
+    const addToInvestigationButton = new AddToInvestigationButton({
+      query: traceId,
+      type: 'trace',
+      dsUid,
+      $data: queryRunner,
+    });
+    
+    addToInvestigationButton.activate();
+    this.setState({ addToInvestigationButton });
+    this._subs.add(
+      addToInvestigationButton.subscribeToState(() => {
+        this.updateInvestigationLink();
+      })
+    );
+        
+    queryRunner.activate();
+    
+    this._subs.add(
+      queryRunner.subscribeToState((state) => {
+        if (state.data?.state === LoadingState.Done && state.data?.series?.length > 0) {
+          const serviceNameField = state.data.series[0]?.fields?.find((f) => f.name === 'serviceName');
+          
+          if (serviceNameField && serviceNameField.values[0]) {
+            addToInvestigationButton.setState({
+              ...addToInvestigationButton.state,
+              labelValue: `${serviceNameField.values[0]}`,
+            });
+          }
+        }
+      })
+    );
+    
+    addToInvestigationButton.setState({
+      ...addToInvestigationButton.state,
+      labelValue: traceId,
+    });
+  }
+
+  private async updateInvestigationLink() {
+    const { addToInvestigationButton } = this.state;
+    if (!addToInvestigationButton) { 
+      return;
+    }
+
+    const link = await getInvestigationLink(addToInvestigationButton);
+    if (link) {
+      this.setState({ investigationLink: link });
+    }
+  }
+
   static Component = ({ model }: SceneComponentProps<TraceExploration>) => {
     const { body } = model.useState();
     const styles = useStyles2(getStyles);
@@ -167,7 +241,7 @@ export class TraceExploration extends SceneObjectBase<TraceExplorationState> {
 export class TraceExplorationScene extends SceneObjectBase {
   static Component = ({ model }: SceneComponentProps<TraceExplorationScene>) => {
     const traceExploration = getTraceExplorationScene(model);
-    const { controls, topScene, drawerScene, traceId, issueDetector } = traceExploration.useState();
+    const { controls, topScene, drawerScene, traceId, issueDetector, investigationLink, addToInvestigationButton } = traceExploration.useState();
     const { hasIssue } = issueDetector?.useState() || {
       hasIssue: false,
     };
@@ -206,6 +280,19 @@ export class TraceExplorationScene extends SceneObjectBase {
         </div>
       </Menu>
     );
+
+    const addToInvestigationClicked = (e: React.MouseEvent) => {
+      if (investigationLink?.onClick) {
+        investigationLink.onClick(e);
+      }
+      
+      reportAppInteraction(
+        USER_EVENTS_PAGES.analyse_traces,
+        USER_EVENTS_ACTIONS.analyse_traces.add_to_investigation_trace_view_clicked
+      );
+      
+      setTimeout(() => traceExploration.closeDrawer(), 100);
+    };
 
     return (
       <>
@@ -254,7 +341,27 @@ export class TraceExplorationScene extends SceneObjectBase {
           <div className={styles.body}>{topScene && <topScene.Component model={topScene} />}</div>
         </div>
         {drawerScene && traceId && (
-          <Drawer title={`View trace ${traceId}`} size={'lg'} onClose={() => traceExploration.closeDrawer()}>
+          <Drawer size={'lg'} onClose={() => traceExploration.closeDrawer()}>
+            <div className={styles.drawerHeader}>
+              <h3>View trace {traceId}</h3>
+                <div className={styles.drawerHeaderButtons}>
+                  {addToInvestigationButton && investigationLink && (
+                    <Button
+                      variant='secondary'
+                      size='sm'
+                      icon='plus-square'
+                      onClick={addToInvestigationClicked}
+                    >
+                      {ADD_TO_INVESTIGATION_MENU_TEXT}
+                    </Button>
+                  )}
+                  <IconButton 
+                    name='times' 
+                    onClick={() => traceExploration.closeDrawer()} 
+                    tooltip='Close drawer'
+                  />
+                </div>
+            </div>
             <drawerScene.Component model={drawerScene} />
           </Drawer>
         )}
@@ -345,6 +452,20 @@ function getStyles(theme: GrafanaTheme2) {
       padding: `0 ${theme.spacing(2)} ${theme.spacing(2)} ${theme.spacing(2)}`,
       overflow: 'auto', /* Needed for sticky positioning */
       maxHeight: '100%' /* Needed for sticky positioning */
+    }),
+    drawerHeader: css({
+      display: 'flex',
+      justifyContent: 'space-between',
+      alignItems: 'center',
+      borderBottom: `1px solid ${theme.colors.border.weak}`,
+      paddingBottom: theme.spacing(2),
+      marginBottom: theme.spacing(2),
+    }),
+    drawerHeaderButtons: css({
+      display: 'flex',
+      justifyContent: 'flex-end',
+      gap: theme.spacing(1.5),
+      padding: `0 0 ${theme.spacing(1)} 0`,
     }),
     body: css({
       label: 'body',
